pyproject.toml
[project]
name = "llm-stock-analyzer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "fastapi[standard]>=0.115.13",
    "html5lib>=1.1",
    "langchain-openai>=0.3.27",
    "loguru>=0.7.3",
    "lxml>=6.0.0",
    "pandas>=2.3.0",
    "pydantic>=2.11.5",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "shioaji>=1.2.6",
    "uvicorn>=0.34.3",
    "yfinance>=0.2.64",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "ruff>=0.12.0",
    "pyright>=1.1.402",
]
